services/labs-service/app/requirements.txt
grpcio==1.67.1
grpcio-tools==1.67.1
sqlalchemy==2.0.29
sqlalchemy-serializer==1.4.22
python-dotenv==1.0.1
psycopg2-binary==2.9.9
minio==7.2.10
pytest==8.4.1
pymongo==4.13.2
grpcio-health-checking 